% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_member.R
\name{add_team_member}
\alias{add_team_member}
\title{Add your team member}
\usage{
add_team_member(gitstats_obj, member_name, gh_login = NULL, gl_login = NULL)
}
\arguments{
\item{gitstats_obj}{\code{GitStats} object.}

\item{member_name}{Name of a member.}

\item{gh_login}{A GitHub login.}

\item{gl_login}{A GitLab login.}
}
\value{
\code{GitStats} object with new information on team member.
}
\description{
Passes information on team member to your \code{team} field.
}
