% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{check_package_usage}
\alias{check_package_usage}
\title{Check package usage across repositories}
\usage{
check_package_usage(gitstats_obj, package_name)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{package_name}{A character, name of the package.}
}
\value{
A table of repositories content.
}
\description{
Wrapper over searching repositories by code blobs related to
using package (\code{library(package)}, \code{require(package)} and \verb{package::}).
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  check_package_usage("Shiny")
}
}
