% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{create_gitstats}
\alias{create_gitstats}
\title{Create a \code{GitStats} object}
\usage{
create_gitstats()
}
\value{
A \code{GitStats} object.
}
\description{
Create a \code{GitStats} object
}
\examples{
my_gitstats <- create_gitstats()
}
