% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineGraphQL.R
\name{EngineGraphQL}
\alias{EngineGraphQL}
\title{A EngineGraphQL class}
\description{
A superclass for methods wraping GraphQL API responses.
}
\section{Super class}{
\code{\link[GitStats:Engine]{GitStats::Engine}} -> \code{EngineGraphQL}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{git_service}}{A character specifying whether GitHub or GitLab.}

\item{\code{gql_api_url}}{A character, url of GraphQL API.}

\item{\code{gql_query}}{An environment for GraphQL queries.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineGraphQL$new()}}
\item \href{#method-gql_response}{\code{EngineGraphQL$gql_response()}}
\item \href{#method-clone}{\code{EngineGraphQL$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GraphQL} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$new(token = NA, gql_api_url = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{A token.}

\item{\code{gql_api_url}}{A character, url of GraphQL API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{GraphQL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gql_response"></a>}}
\if{latex}{\out{\hypertarget{method-gql_response}{}}}
\subsection{Method \code{gql_response()}}{
Wrapper of GraphQL API request and response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$gql_response(gql_query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_query}}{A string with GraphQL query.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
