% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineGraphQLGitHub.R
\name{EngineGraphQLGitHub}
\alias{EngineGraphQLGitHub}
\title{A EngineGraphQLGitHub class}
\description{
A class for methods wrapping GitHub's GraphQL API responses.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gql_api_url}}{A character, url of GraphQL API.}

\item{\code{gql_query}}{An environment for GraphQL queries.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineGraphQLGitHub$new()}}
\item \href{#method-gql_response}{\code{EngineGraphQLGitHub$gql_response()}}
\item \href{#method-get_repos}{\code{EngineGraphQLGitHub$get_repos()}}
\item \href{#method-get_commits}{\code{EngineGraphQLGitHub$get_commits()}}
\item \href{#method-clone}{\code{EngineGraphQLGitHub$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create \code{EngineGraphQLGitHub} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitHub$new(gql_api_url, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_api_url}}{GraphQL API url.}

\item{\code{token}}{A token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gql_response"></a>}}
\if{latex}{\out{\hypertarget{method-gql_response}{}}}
\subsection{Method \code{gql_response()}}{
Wrapper of GraphQL API request and response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitHub$gql_response(gql_query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_query}}{A string with GraphQL query.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to retrieve all repositories for an organization in
a table format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitHub$get_repos(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
Method to pull all commits from organization, optionally
filtered by team members.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitHub$get_commits(org, date_from, date_until, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{date_from}}{A starting date to look commits for.}

\item{\code{date_until}}{An end date to look commits for.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of commits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitHub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
