% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_commits}
\alias{get_commits}
\title{Get commits}
\usage{
get_commits(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of commits.
}
\description{
Retrieves commits table pulled by \code{GitStats}.
}
\examples{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_commits(
    date_from = "2023-01-01",
    date_until = "2023-01-15"
  ) \%>\%
  get_commits()
}
