% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_commits_stats}
\alias{get_commits_stats}
\title{Get statistics on commits}
\usage{
get_commits_stats(gitstats_obj, time_interval = c("month", "day", "week"))
}
\arguments{
\item{gitstats_obj}{A GitStats class object.}

\item{time_interval}{A character, specifying time interval to show statistics.}
}
\value{
A table of \code{commits_stats} class.
}
\description{
Prepare statistics from the pulled commits data.
}
\examples{
 create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_commits(
    date_from = "2023-03-01",
    date_until = "2023-03-15"
  ) \%>\%
  get_commits_stats()
}
