% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_files}
\alias{get_files}
\title{Get files}
\usage{
get_files(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of files content.
}
\description{
Retrieves files table pulled by \code{GitStats}.
}
