% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_files_content}
\alias{get_files_content}
\title{Get content of files}
\usage{
get_files_content(
  gitstats_object,
  file_path,
  cache = TRUE,
  verbose = is_verbose(gitstats_object)
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{file_path}{A standardized path to file(s) in repositories. May be a
character vector if multiple files are to be pulled.}

\item{cache}{A logical, if set to \code{TRUE} GitStats will retrieve the last
result from its storage.}

\item{verbose}{A logical, \code{TRUE} by default. If \code{FALSE} messages and printing
output is switched off.}
}
\value{
A data.frame.
}
\description{
Pull text files content for a given scope (orgs, repos or whole
git hosts).
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_github_host(
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_gitlab_host(
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  )
 get_files(
   gitstats_obj = my_gitstats,
   file_path = c("LICENSE", "DESCRIPTION"))
}
}
