% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_orgs}
\alias{get_orgs}
\title{Get organizations}
\usage{
get_orgs(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A vector of organizations.
}
\description{
Retrieves organizations set or pulled by \code{GitStats}. Especially
helpful when user is scanning whole git platform and wants to have a
glimpse at organizations.
}
\examples{
create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  get_orgs()
}
