% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pull_requests.R
\name{get_pull_requests_stats}
\alias{get_pull_requests_stats}
\title{Get pull requests statistics}
\usage{
get_pull_requests_stats(
  pull_requests,
  time_aggregation = c("year", "month", "week", "day"),
  group_var
)
}
\arguments{
\item{pull_requests}{A \code{gitstats_pr_stats} S3 class table object (output of \code{get_pull_requests()}).}
}
\value{
A table of \code{pull_requests_stats} class.
}
\description{
Prepare statistics from the pulled pull requests data.
}
\details{
To make function work, you need first to get pull requests data with
\code{GitStats}. See examples section.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() |>
   set_github_host(
     token = Sys.getenv("GITHUB_PAT"),
     repos = c("r-world-devs/GitStats", "openpharma/visR")
   ) |>
   get_pull_requests(my_gitstats, since = "2022-01-01") |>
   get_pull_requests_stats(
     time_aggregation = "month",
     group_var = author
   )
}
}
