% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_R_package_usage}
\alias{get_R_package_usage}
\title{Get information on package usage across repositories}
\usage{
get_R_package_usage(
  gitstats_object,
  package_name,
  only_loading = FALSE,
  use_storage = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{package_name}{A character, name of the package.}

\item{only_loading}{A boolean, if \code{TRUE} function will check only if package
is loaded in repositories, not used as dependencies.}

\item{use_storage}{A logical, should storage be used.}

\item{verbose}{A logical, shoud verbose mode be used.}
}
\value{
A table of repositories content.
}
\description{
Wrapper over searching repositories by code blobs related to
loading package (\code{library(package)} and \code{require(package)} in all files) or
using it as a dependency (\code{package} in \code{DESCRIPTION} and \code{NAMESPACE} files).
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  get_R_package_usage("shiny")
}
}
