% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_release_logs}
\alias{get_release_logs}
\title{Get release logs}
\usage{
get_release_logs(gitstats_obj, since = NULL, until = NULL)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{since}{A starting date for release logs.}

\item{until}{An end date for release logs.}
}
\value{
A table with release logs.
}
\description{
Pull release logs from repositories.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  )
  get_release_logs(my_gistats, since = "2024-01-01")
}
}
