% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{get_repos}
\alias{get_repos}
\title{Get information on repositories}
\usage{
get_repos(
  gitstats_obj,
  by = "org",
  phrase = NULL,
  language = NULL,
  print_out = TRUE
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{by}{A character, to choose between: \itemize{ \item{org}{Organizations
\itemize{
\item owners of repositories} \item(team){A team} \item(phrase){A keyword in
code blobs.}}
}}

\item{phrase}{A phrase to look for in codelines.}

\item{language}{A character specifying language used in repositories.}

\item{print_out}{A boolean stating whether to print an output.}
}
\value{
A data.frame of repositories
}
\description{
List all repositories for an organization, a team or by a
codephrase.
}
