% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_repos_contributors}
\alias{get_repos_contributors}
\title{Add information on contributors to your repositories.}
\usage{
get_repos_contributors(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of repositories with added information on contributors.
}
\description{
A method to add information on repository contributors.
}
