% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_repos_stats}
\alias{get_repos_stats}
\title{Get statistics on repositories}
\usage{
get_repos_stats(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats class object.}
}
\value{
A table of \code{repos_stats} class.
}
\description{
Prepare statistics from the pulled repositories data.
}
\examples{
 create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_repos() \%>\%
  get_repos_stats()
}
