% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_storage}
\alias{get_storage}
\title{Get data from \code{GitStats} storage}
\usage{
get_storage(gitstats_object, storage = NULL)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{storage}{A character, type of the data you want to get from storage:
\code{commits}, \code{repositories}, \code{release_logs}, \code{users}, \code{files},
\code{files_structure}, \code{R_package_usage} or \code{release_logs}.}
}
\value{
A list of tibbles (if \code{storage} set to \code{NULL}) or a tibble (if
\code{storage} defined).
}
\description{
Retrieves whole or particular data (see \code{storage} parameter)
pulled earlier with \code{GitStats}.
}
