% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_table}
\alias{get_table}
\title{Get data}
\usage{
get_table(gitstats_obj, table)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{table}{Type of table (\code{repositories}, \code{commits}, \code{files}, \code{users},
\code{R_package_usage}, \code{release_logs}).}
}
\value{
A table.
}
\description{
Retrieves data pulled by \code{GitStats}.
}
