% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_users}
\alias{get_users}
\title{Get users}
\usage{
get_users(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of users.
}
\description{
Retrieves users table pulled by \code{GitStats}.
}
\examples{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_users(c("maciekbanas", "marcinkowskak")) \%>\%
  get_users()
}
