% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_github_enterprise.R
\name{GitHubEnterpriseClient}
\alias{GitHubEnterpriseClient}
\title{A GitHub Enterprise API Client class}
\description{
An object with methods to derive information form GitHub Enterprise API.
}
\section{Super classes}{
\code{\link[GitStats:GitClient]{GitStats::GitClient}} -> \code{\link[GitStats:GitHubClient]{GitStats::GitHubClient}} -> \code{GitHubEnterpriseClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_commits_by_team}{\code{GitHubEnterpriseClient$get_commits_by_team()}}
\item \href{#method-clone}{\code{GitHubEnterpriseClient$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="get_commits_by_owner">}\href{../../GitStats/html/GitHubClient.html#method-get_commits_by_owner}{\code{GitStats::GitHubClient$get_commits_by_owner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="get_repos_by_codephrase">}\href{../../GitStats/html/GitHubClient.html#method-get_repos_by_codephrase}{\code{GitStats::GitHubClient$get_repos_by_codephrase()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="get_repos_by_owner">}\href{../../GitStats/html/GitHubClient.html#method-get_repos_by_owner}{\code{GitStats::GitHubClient$get_repos_by_owner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="get_repos_by_team">}\href{../../GitStats/html/GitHubClient.html#method-get_repos_by_team}{\code{GitStats::GitHubClient$get_repos_by_team()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="initialize">}\href{../../GitStats/html/GitHubClient.html#method-initialize}{\code{GitStats::GitHubClient$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitHubClient" data-id="print">}\href{../../GitStats/html/GitHubClient.html#method-print}{\code{GitStats::GitHubClient$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_by_team}{}}}
\subsection{Method \code{get_commits_by_team()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHubEnterpriseClient$get_commits_by_team(
  team,
  owners = self$owners,
  date_from,
  date_until = Sys.time()
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHubEnterpriseClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
