% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gitlab.R
\name{GitLab}
\alias{GitLab}
\title{A GitLab API Client class}
\description{
An object with methods to derive information form GitLab API.
}
\section{Super class}{
\code{\link[GitStats:GitService]{GitStats::GitService}} -> \code{GitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_repos_by_org}{\code{GitLab$get_repos_by_org()}}
\item \href{#method-get_repos_by_team}{\code{GitLab$get_repos_by_team()}}
\item \href{#method-get_repos_by_codephrase}{\code{GitLab$get_repos_by_codephrase()}}
\item \href{#method-get_commits_by_org}{\code{GitLab$get_commits_by_org()}}
\item \href{#method-get_commits_by_team}{\code{GitLab$get_commits_by_team()}}
\item \href{#method-print}{\code{GitLab$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitService" data-id="initialize">}\href{../../GitStats/html/GitService.html#method-initialize}{\code{GitStats::GitService$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_org"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_org}{}}}
\subsection{Method \code{get_repos_by_org()}}{
A method to list all repositories for an organization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_repos_by_org(orgs = self$orgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orgs}}{A character vector of organisations (project groups).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_team}{}}}
\subsection{Method \code{get_repos_by_team()}}{
A method to list all repositories
for a team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_repos_by_team(team, orgs = self$orgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A list of team members. Specified
by \code{set_team()} method of GitStats class
object.}

\item{\code{orgs}}{A character vector of organisations (project groups).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_codephrase"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_codephrase}{}}}
\subsection{Method \code{get_repos_by_codephrase()}}{
A method to find repositories with given phrase in code lines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_repos_by_codephrase(phrase, language = "R")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phrase}}{A phrase to look for in codelines.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_by_org"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_by_org}{}}}
\subsection{Method \code{get_commits_by_org()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_commits_by_org(orgs = self$orgs, date_from, date_until = Sys.time())}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_by_team}{}}}
\subsection{Method \code{get_commits_by_team()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_commits_by_team(
  team,
  orgs = self$orgs,
  date_from,
  date_until = Sys.time()
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitLab object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$print()}\if{html}{\out{</div>}}
}

}
}
