% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gitlab.R
\name{GitLabClient}
\alias{GitLabClient}
\title{A GitLab API Client class}
\description{
An object with methods to derive information form GitLab API.
}
\section{Super class}{
\code{\link[GitStats:GitClient]{GitStats::GitClient}} -> \code{GitLabClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitLabClient$new()}}
\item \href{#method-get_projects_by_group}{\code{GitLabClient$get_projects_by_group()}}
\item \href{#method-get_repos_by_team}{\code{GitLabClient$get_repos_by_team()}}
\item \href{#method-get_repos_by_codephrase}{\code{GitLabClient$get_repos_by_codephrase()}}
\item \href{#method-get_commits_by_group}{\code{GitLabClient$get_commits_by_group()}}
\item \href{#method-get_commits_by_team}{\code{GitLabClient$get_commits_by_team()}}
\item \href{#method-print}{\code{GitLabClient$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitLabClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$new(rest_api_url = NA, token = NA, groups = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rest_api_url}}{A url of rest API.}

\item{\code{token}}{A token.}

\item{\code{groups}}{A character vector of groups of projects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GitLabClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_projects_by_group"></a>}}
\if{latex}{\out{\hypertarget{method-get_projects_by_group}{}}}
\subsection{Method \code{get_projects_by_group()}}{
A method to list all repositories for an organization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$get_projects_by_group(project_groups = self$groups)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_groups}}{A character vector of project groups.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_team}{}}}
\subsection{Method \code{get_repos_by_team()}}{
A method to list all repositories
for a team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$get_repos_by_team(team, project_groups = self$groups)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A list of team members. Specified
by \code{set_team()} method of GitStats class
object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_codephrase"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_codephrase}{}}}
\subsection{Method \code{get_repos_by_codephrase()}}{
A method to find repositories with given phrase in code lines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$get_repos_by_codephrase(phrase, language = "R")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phrase}}{A phrase to look for in codelines.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_by_group"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_by_group}{}}}
\subsection{Method \code{get_commits_by_group()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$get_commits_by_group(
  groups = self$groups,
  date_from,
  date_until = Sys.time()
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_by_team}{}}}
\subsection{Method \code{get_commits_by_team()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$get_commits_by_team(
  team,
  groups = self$groups,
  date_from,
  date_until = Sys.time(),
  by
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitHubClient object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLabClient$print()}\if{html}{\out{</div>}}
}

}
}
