% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitPlatform.R
\name{GitPlatform}
\alias{GitPlatform}
\title{A Git Service API Client superclass}
\description{
A superclass for GitHub and GitLab classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{graphql_engine}}{A GraphQL engine.}

\item{\code{rest_engine}}{A REST engine.}

\item{\code{orgs}}{A character vector of organizations.}

\item{\code{git_service}}{A character specifying whether GitHub or GitLab.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitPlatform$new()}}
\item \href{#method-get_repos}{\code{GitPlatform$get_repos()}}
\item \href{#method-clone}{\code{GitPlatform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitPlatform} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitPlatform$new(orgs = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GitPlatform} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to list all repositories for an organization, a
team or by a keyword.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitPlatform$get_repos(by, team, phrase, language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{A character, to choose between: \itemize{\item{org -
organizations (project groups)} \item{team -
A team} \item{phrase - A keyword in code blobs.}}}

\item{\code{team}}{A list of team members.}

\item{\code{phrase}}{A character to look for in code blobs. Obligatory if
\code{by} parameter set to \code{"phrase"}.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitPlatform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
