% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitService.R
\name{GitService}
\alias{GitService}
\title{A Git Service API Client superclass}
\description{
A superclass for GitHub and GitLab classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rest_api_url}}{A character, url of REST API.}

\item{\code{gql_api_url}}{A character, url of GraphQL API.}

\item{\code{orgs}}{A character vector of organizations.}

\item{\code{enterprise}}{A boolean defining whether Git Service is public or
enterprise version.}

\item{\code{token}}{A token authorizing access to API.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{token}}{A token authorizing access to API.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitService$new()}}
\item \href{#method-clone}{\code{GitService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitService} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitService$new(rest_api_url = NA, gql_api_url = NA, token = NA, orgs = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rest_api_url}}{A url of rest API.}

\item{\code{gql_api_url}}{A url of GraphQL API.}

\item{\code{token}}{A token.}

\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GitService} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
