% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{GitStats}
\alias{GitStats}
\title{A statistics platform for Git hosts}
\description{
An R6 class object with methods to derive information from multiple Git platforms.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_params}{\code{GitStats$set_params()}}
\item \href{#method-set_host}{\code{GitStats$set_host()}}
\item \href{#method-set_team_member}{\code{GitStats$set_team_member()}}
\item \href{#method-pull_repos}{\code{GitStats$pull_repos()}}
\item \href{#method-pull_repos_contributors}{\code{GitStats$pull_repos_contributors()}}
\item \href{#method-pull_commits}{\code{GitStats$pull_commits()}}
\item \href{#method-pull_users}{\code{GitStats$pull_users()}}
\item \href{#method-get_orgs}{\code{GitStats$get_orgs()}}
\item \href{#method-get_repos}{\code{GitStats$get_repos()}}
\item \href{#method-get_commits}{\code{GitStats$get_commits()}}
\item \href{#method-get_users}{\code{GitStats$get_users()}}
\item \href{#method-print}{\code{GitStats$print()}}
\item \href{#method-clone}{\code{GitStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
Set up your search settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_params(search_param, team_name, phrase, language, print_out)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_param}}{One of three: \code{team}, \code{org} or \code{phrase}.}

\item{\code{team_name}}{A name of a team.}

\item{\code{phrase}}{A phrase to look for.}

\item{\code{language}}{A language of programming code.}

\item{\code{print_out}}{A boolean stating if you want to print output after
pulling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_host"></a>}}
\if{latex}{\out{\hypertarget{method-set_host}{}}}
\subsection{Method \code{set_host()}}{
Method to set connections to Git platforms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_host(api_url, token, orgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_url}}{A character, url address of API.}

\item{\code{token}}{A token.}

\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, puts connection information into \verb{$hosts} slot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_team_member"></a>}}
\if{latex}{\out{\hypertarget{method-set_team_member}{}}}
\subsection{Method \code{set_team_member()}}{
A method to add a team member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_team_member(member_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{member_name}}{Name of a member.}

\item{\code{...}}{User names on Git platforms.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, pass information on team member to \code{GitStats}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_repos"></a>}}
\if{latex}{\out{\hypertarget{method-pull_repos}{}}}
\subsection{Method \code{pull_repos()}}{
A method to list all repositories for an organization,
a team or by a keyword.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$pull_repos(add_contributors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_contributors}}{A boolean to decide whether to add contributors
information to repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_repos_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-pull_repos_contributors}{}}}
\subsection{Method \code{pull_repos_contributors()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$pull_repos_contributors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_commits"></a>}}
\if{latex}{\out{\hypertarget{method-pull_commits}{}}}
\subsection{Method \code{pull_commits()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$pull_commits(date_from, date_until)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{A starting date to look commits for}

\item{\code{date_until}}{An end date to look commits for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of commits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_users"></a>}}
\if{latex}{\out{\hypertarget{method-pull_users}{}}}
\subsection{Method \code{pull_users()}}{
Get information on users.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$pull_users(users)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{users}}{Character vector of users.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of users.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_orgs"></a>}}
\if{latex}{\out{\hypertarget{method-get_orgs}{}}}
\subsection{Method \code{get_orgs()}}{
Print organizations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_orgs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
Print repositories output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
Print commits output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_commits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_users"></a>}}
\if{latex}{\out{\hypertarget{method-get_users}{}}}
\subsection{Method \code{get_users()}}{
Print users output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_users()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitStats object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
