% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gitstats_plot}
\alias{gitstats_plot}
\title{Plot Git statistics}
\usage{
gitstats_plot(
  stats_table = NULL,
  value_to_plot = NULL,
  plotly_mode = FALSE,
  value_decreasing = TRUE,
  n = NULL
)
}
\arguments{
\item{stats_table}{A table with repository or commits statistics.}

\item{value_to_plot}{Value to be plotted.}

\item{plotly_mode}{A boolean, if TRUE, turns plot into interactive plotly.}

\item{value_decreasing}{A boolean to set ordering of a value on the plot.}

\item{n}{An integer, a maximum number of repos/organizations to show on the plot.}
}
\value{
A plot.
}
\description{
A generic to plot statistics from repositories or commits.
}
