% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GQLQueryGitHub.R
\name{GQLQueryGitHub}
\alias{GQLQueryGitHub}
\title{A GQLQueryGitHub class}
\description{
A class with methods to build GraphQL Queries for GitHub.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-orgs}{\code{GQLQueryGitHub$orgs()}}
\item \href{#method-repos_by_org}{\code{GQLQueryGitHub$repos_by_org()}}
\item \href{#method-repos_by_user}{\code{GQLQueryGitHub$repos_by_user()}}
\item \href{#method-user}{\code{GQLQueryGitHub$user()}}
\item \href{#method-commits_by_repo}{\code{GQLQueryGitHub$commits_by_repo()}}
\item \href{#method-clone}{\code{GQLQueryGitHub$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orgs"></a>}}
\if{latex}{\out{\hypertarget{method-orgs}{}}}
\subsection{Method \code{orgs()}}{
Prepare query to list organizations from GitHub.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$orgs(end_cursor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end_cursor}}{An end cursor to paginate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repos_by_org"></a>}}
\if{latex}{\out{\hypertarget{method-repos_by_org}{}}}
\subsection{Method \code{repos_by_org()}}{
Prepare query to get repositories from GitHub.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$repos_by_org(repo_cursor = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo_cursor}}{An end cursor for repositories page.}

\item{\code{org}}{An organization of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repos_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-repos_by_user}{}}}
\subsection{Method \code{repos_by_user()}}{
Prepare query to get repositories from GitHub.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$repos_by_user(repo_cursor = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo_cursor}}{An end cursor for repositories page.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user"></a>}}
\if{latex}{\out{\hypertarget{method-user}{}}}
\subsection{Method \code{user()}}{
Prepare query to get info on a GitHub user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$user(login)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{login}}{A login of a user.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commits_by_repo"></a>}}
\if{latex}{\out{\hypertarget{method-commits_by_repo}{}}}
\subsection{Method \code{commits_by_repo()}}{
Prepare query to get commits on GitHub.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$commits_by_repo(
  org,
  repo,
  since,
  until,
  commits_cursor = "",
  author_id = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A GitHub organization.}

\item{\code{repo}}{Name of a repository.}

\item{\code{since}}{Git Time Stamp of starting date of commits.}

\item{\code{until}}{Git Time Stamp of end date of commits.}

\item{\code{commits_cursor}}{An endCursor.}

\item{\code{author_id}}{An Id of an author.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitHub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
