% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQL.R
\name{GraphQL}
\alias{GraphQL}
\title{A GraphQL class}
\description{
An object with methods to build GraphQL Queries.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gh_repos_by_user}{\code{GraphQL$gh_repos_by_user()}}
\item \href{#method-users_id}{\code{GraphQL$users_id()}}
\item \href{#method-gh_commits}{\code{GraphQL$gh_commits()}}
\item \href{#method-gl_groups_by_user}{\code{GraphQL$gl_groups_by_user()}}
\item \href{#method-clone}{\code{GraphQL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gh_repos_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-gh_repos_by_user}{}}}
\subsection{Method \code{gh_repos_by_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQL$gh_repos_by_user(user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-users_id"></a>}}
\if{latex}{\out{\hypertarget{method-users_id}{}}}
\subsection{Method \code{users_id()}}{
Gets ID of a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQL$users_id(login)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{login}}{A login of a user.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gh_commits"></a>}}
\if{latex}{\out{\hypertarget{method-gh_commits}{}}}
\subsection{Method \code{gh_commits()}}{
Query to get commits on GitHub
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQL$gh_commits(org, repo, since, until, cursor = "", author_id = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A GitHub organization}

\item{\code{repo}}{Name of a repository}

\item{\code{since}}{Git Time Stamp of starting date of commits.}

\item{\code{until}}{Git Time STamp of end date of commits.}

\item{\code{cursor}}{An endCursor.}

\item{\code{author_id}}{An Id of an author.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gl_groups_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-gl_groups_by_user}{}}}
\subsection{Method \code{gl_groups_by_user()}}{
GitLab. Method to build query to pull groups by users.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQL$gl_groups_by_user(team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A string of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
