% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQLQuery.R
\name{GraphQLQuery}
\alias{GraphQLQuery}
\title{A GraphQLQuery class}
\description{
An object with methods to build GraphQL Queries.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-groups_by_user}{\code{GraphQLQuery$groups_by_user()}}
\item \href{#method-clone}{\code{GraphQLQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-groups_by_user}{}}}
\subsection{Method \code{groups_by_user()}}{
Method to build query to pull groups by users.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$groups_by_user(team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A string of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
