% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQLQueryGitLab.R
\name{GraphQLQueryGitLab}
\alias{GraphQLQueryGitLab}
\title{A GraphQLQueryGitLab class}
\description{
A class with methods to build GraphQL Queries for GitLab.
}
\section{Super class}{
\code{\link[GitStats:GraphQLQuery]{GitStats::GraphQLQuery}} -> \code{GraphQLQueryGitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-groups_by_user}{\code{GraphQLQueryGitLab$groups_by_user()}}
\item \href{#method-projects_by_group}{\code{GraphQLQueryGitLab$projects_by_group()}}
\item \href{#method-clone}{\code{GraphQLQueryGitLab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-groups_by_user}{}}}
\subsection{Method \code{groups_by_user()}}{
Method to build query to pull GitLab groups by users.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQueryGitLab$groups_by_user(team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A string of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projects_by_group"></a>}}
\if{latex}{\out{\hypertarget{method-projects_by_group}{}}}
\subsection{Method \code{projects_by_group()}}{
Method to build query to pull projects by group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQueryGitLab$projects_by_group(group, projects_cursor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{A group of projects.}

\item{\code{projects_cursor}}{A cursor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQueryGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
