% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{is_parallel}
\alias{is_parallel}
\title{Check if parallel processing is active}
\usage{
is_parallel(gitstats)
}
\arguments{
\item{gitstats}{A GitStats object.}
}
\value{
A logical scalar.
}
\description{
Returns \code{TRUE} when mirai daemons are running on the given
\code{GitStats} object (i.e. \code{set_parallel()} has been called), \code{FALSE}
otherwise.
}
\examples{
\dontrun{
  my_gitstats <- create_gitstats()
  is_parallel(my_gitstats)
}
}
