% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commits.R
\name{plot_commits_stats}
\alias{plot_commits_stats}
\alias{plot_commit_lines}
\title{Plot commits additions and deletions.}
\usage{
plot_commit_lines(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A plot.
}
\description{
Plot commits additions and deletions.
}
