% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_repos}
\alias{plot_repos}
\title{Plot repositories statistics}
\usage{
plot_repos(
  repos_data = NULL,
  value_to_plot = "last_activity",
  value_decreasing = TRUE,
  plotly_mode = FALSE,
  n = 10
)
}
\arguments{
\item{repos_data}{A table with repositories data.}

\item{value_to_plot}{Value to be plotted.}

\item{value_decreasing}{A boolean to set ordering of a value on the plot.}

\item{plotly_mode}{A boolean, if TRUE, turns plot into interactive plotly.}

\item{n}{An integer, a maximum number of repos/organizations to show on the plot.}
}
\value{
A plot.
}
\description{
A function to plot statistics from repositories.
}
