% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_commits}
\alias{pull_commits}
\title{Pull information on commits}
\usage{
pull_commits(gitstats_obj, date_from = NULL, date_until = Sys.time())
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{date_from}{A starting date of commits.}

\item{date_until}{An end date of commits.}
}
\value{
A \code{GitStats} class object with commits table.
}
\description{
List all commits from all repositories for an organization or a
team.
}
\examples{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  set_params(
    search_param = "team",
    team_name = "rwdevs"
  ) \%>\%
  set_team_member("Maciej Banaś", "maciekbanas") \%>\%
  pull_commits(
    date_from = "2023-01-01",
    date_until = "2023-01-15"
  )
}
