% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_files}
\alias{pull_files}
\title{Pull files content}
\usage{
pull_files(gitstats_obj, file_path)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{file_path}{A standardized path to file(s) in repositories. May be a
character vector if multiple files are to be pulled.}
}
\value{
A \code{GitStats} object with table of files.
}
\description{
Pull files content from Git Hosts.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_files("meta_data.yaml")
}
}
