% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_R_package_usage}
\alias{pull_R_package_usage}
\title{Check package usage across repositories}
\usage{
pull_R_package_usage(gitstats_obj, package_name, only_loading = FALSE)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{package_name}{A character, name of the package.}

\item{only_loading}{A boolean, if \code{TRUE} function will check only if package
is loaded in repositories, not used as dependencies. This is much faster
approach as searching usage only with loading (i.e. library(package)) is
based on Search APIs (one endpoint), whereas searching usage as a
dependency pulls text files from all repositories (many endpoints). This is
a good option to choose when you want to check package usage but guess that
it may be used mainly by loading in data scripts and not used as a
dependency of other packages.}
}
\value{
A table of repositories content.
}
\description{
Wrapper over searching repositories by code blobs related to
using package (\code{library(package)}, \code{require(package)} and \verb{package::}).
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  pull_R_package_usage("Shiny")
}
}
