% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_release_logs}
\alias{pull_release_logs}
\title{Pull release logs}
\usage{
pull_release_logs(gitstats_obj, date_from, date_until = Sys.Date())
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{date_from}{A starting date for release logs.}

\item{date_until}{An end date for release logs.}
}
\value{
A table with release logs.
}
\description{
Pull release logs from repositories.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  pull_release_logs(date_from = "2024-01-01")
}
}
