% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_repos_contributors}
\alias{pull_repos_contributors}
\title{Pull information on contributors}
\usage{
pull_repos_contributors(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} class object with repositories table with added
information on contributors.
}
\description{
Adds information on contributors to already pulled repositories
table.
}
