% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_users}
\alias{pull_users}
\title{Pull users data}
\usage{
pull_users(gitstats_obj, users)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{users}{A character vector of users.}
}
\value{
A table of users.
}
\description{
Pull users data from Git Host.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_users(c("maciekabanas", "marcinkowskak"))
}
}
