% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{remove_from_storage}
\alias{remove_from_storage}
\title{Remove a table from \code{GitStats} storage}
\usage{
remove_from_storage(gitstats, storage)
}
\arguments{
\item{gitstats}{A GitStats object.}

\item{storage}{A character, name of the table to remove (e.g. \code{"commits"},
\code{"repositories"}).}
}
\value{
A \code{GitStats} object (invisibly).
}
\description{
Removes a named table from the active storage backend.
}
\examples{
\dontrun{
  my_gitstats <- create_gitstats() |>
    set_github_host(
      token = Sys.getenv("GITHUB_PAT"),
      orgs = "r-world-devs"
    )
  get_commits(my_gitstats, since = "2024-01-01")
  remove_from_storage(my_gitstats, storage = "commits")
}
}
