% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{set_host}
\alias{set_host}
\title{Add Git host}
\usage{
set_host(gitstats_obj, api_url, token = NULL, orgs = NULL)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{api_url}{A character, url address of API.}

\item{token}{A token.}

\item{orgs}{A character vector of organisations (owners of repositories in
case of GitHub and groups of projects in case of GitLab). You do not need
to define \code{orgs} if you wish to scan whole internal Git platform (such as
enterprise version of GitHub or GitLab). In case of a public one (like
GitHub) you need to define \code{orgs} as scanning through all organizations
would be an overkill.}
}
\value{
A \code{GitStats} class object with added information on connection
(\verb{$hosts} field).
}
\description{
Add Git host
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "erasmusmc-public-health"
  )
}
}
