% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{set_local_storage}
\alias{set_local_storage}
\title{Set local (in-memory) storage}
\usage{
set_local_storage(gitstats)
}
\arguments{
\item{gitstats}{A GitStats object.}
}
\value{
A \code{GitStats} object (invisibly).
}
\description{
Reset storage to the default in-memory backend.
}
\examples{
\dontrun{
  my_gitstats <- create_gitstats() |>
    set_local_storage()
}
}
