% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_organizations.R
\name{set_organizations}
\alias{set_organizations}
\title{Choose your organizations.}
\usage{
set_organizations(gitstats_obj, ..., api_url = NULL)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{...}{A character vector of oganizations (repo owners or project
groups).}

\item{api_url}{A url for connection.}
}
\value{
A \code{GitStats} class object with added information on organizations
into \verb{$orgs} of a \code{client}.
}
\description{
A method to set you organizations.
}
\examples{
\dontrun{
### with one connection:
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  )

my_gitstats \%>\%
  set_organizations("openpharma", "pharmaverse")

### with multiple connections - remember to specify
`api_url`, when setting organizations:

my_gitstats <- create_gitstats() \%>\%
 set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
 ) \%>\%
 set_connection(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "erasmusmc-public-health"
 )

 my_gitstats \%>\%
  set_organizations(api_url = "https://api.github.com",
                    "openpharma", "pharmaverse")

}
}
