% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{set_params}
\alias{set_params}
\title{Set up your search settings}
\usage{
set_params(
  gitstats_obj,
  search_mode = NULL,
  files = NULL,
  verbose = NULL,
  use_storage = NULL
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{search_mode}{One of four: \code{org}, \code{repo} or \code{code}.}

\item{files}{Define files to scan.}

\item{verbose}{A boolean to decide whether to print output.}

\item{use_storage}{A boolean. If set to \code{TRUE} it will pull data from the
storage when it is there, e.g. \code{repositories}, when user runs
\code{get_repos()}. If set to \code{FALSE} \code{get_repos()} will always pull data from
the API.}
}
\value{
A \code{GitStats} object.
}
\description{
Set up your search settings
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_params(
    search_mode = "code",
    files = "DESCRIPTION"
  )
}
}
