% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{set_team}
\alias{set_team}
\title{Set your team.}
\usage{
set_team(gitstats_obj, team_name)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{team_name}{A name of a team.}
}
\value{
A \code{GitStats} class object with added information to \verb{$team} field.
}
\description{
Declare your team.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_team("RWD-IE")
}
}
