% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_preferences.R
\name{setup_preferences}
\alias{setup_preferences}
\title{Set up your search parameters.}
\usage{
setup_preferences(
  gitstats_obj,
  search_param = NULL,
  team_name = NULL,
  phrase = NULL,
  language = NULL
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{search_param}{One of three: team, orgs or phrase.}

\item{team_name}{Name of a team.}

\item{phrase}{A phrase to look for.}

\item{language}{A language of programming code.}
}
\value{
A \code{GitStats} object.
}
\description{
Set up your search parameters.
}
