% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{show_data}
\alias{show_data}
\title{Show data stored in GitStats}
\usage{
show_data(gitstats_obj, storage)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{storage}{Table of \code{repositories}, \code{commits}, \code{users}, \code{files},
\code{release_logs} or \code{R_package_usage}.}
}
\value{
A table.
}
\description{
Retrieves table stored in \code{GitStats} object.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  get_repos()
  show_data(my_gistats, "repositories")
}
}
