% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{show_hosts}
\alias{show_hosts}
\title{Show hosts set in \code{GitStats}}
\usage{
show_hosts(gitstats)
}
\arguments{
\item{gitstats}{A GitStats object.}
}
\value{
A list of hosts.
}
\description{
Retrieves hosts set by \code{GitStats} with \verb{set_*_host()} functions.
}
