% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{show_orgs}
\alias{show_orgs}
\title{Show organizations}
\usage{
show_orgs(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A vector of organizations.
}
\description{
Prints organizations downloaded in \code{GitStats}. Especially
helpful when user is scanning whole git platform and want to have a glimpse
at organizations.
}
