% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{show_repos}
\alias{show_repos}
\title{Show repositories}
\usage{
show_repos(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of repositories.
}
\description{
Prints repositories table pulled by \code{GitStats}.
}
