% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{show_storage}
\alias{show_storage}
\title{Show content of database.}
\usage{
show_storage(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A data.frame of table names.
}
\description{
Print content of database.
}
