% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_dates}
\alias{standardize_dates}
\title{standardize dates to POSIXct format}
\usage{
standardize_dates(dates)
}
\arguments{
\item{dates}{list of dates}
}
\value{
list of POSIXct dates
}
\description{
standardize dates to POSIXct format
}
