% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{storage_off}
\alias{storage_off}
\title{Switch off storage of data.}
\usage{
storage_off(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} class object with field \verb{$use_storage} turned to FALSE
}
\description{
Switch off storage of data.
}
