% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{storage_on}
\alias{storage_on}
\title{Switch on storage of data.}
\usage{
storage_on(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} class object with field \verb{$use_storage} turned to TRUE
}
\description{
Switch on storage of data.
}
