% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{verbose_off}
\alias{verbose_off}
\title{Switch off verbose mode}
\usage{
verbose_off(gitstats_object)
}
\arguments{
\item{gitstats_object}{A GitStats object.}
}
\value{
A GitStats object.
}
\description{
Stop printing messages and output.
}
