% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{verbose_on}
\alias{verbose_on}
\title{switch on verbose mode}
\usage{
verbose_on(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A GitStats object.
}
\description{
switch on verbose mode
}
