% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineGraphQL.R
\name{EngineGraphQL}
\alias{EngineGraphQL}
\title{A EngineGraphQL class}
\description{
A class for methods wrapping GitHub's GraphQL API responses.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gql_api_url}}{A character, url of GraphQL API.}

\item{\code{gql_query}}{An environment for GraphQL queries.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineGraphQL$new()}}
\item \href{#method-gql_response}{\code{EngineGraphQL$gql_response()}}
\item \href{#method-get_users}{\code{EngineGraphQL$get_users()}}
\item \href{#method-clone}{\code{EngineGraphQL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create \code{EngineGraphQL} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$new(gql_api_url, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_api_url}}{GraphQL API url.}

\item{\code{token}}{A token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gql_response"></a>}}
\if{latex}{\out{\hypertarget{method-gql_response}{}}}
\subsection{Method \code{gql_response()}}{
Wrapper of GraphQL API request and response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$gql_response(gql_query, vars = "null")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_query}}{A string with GraphQL query.}

\item{\code{vars}}{A list of named variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_users"></a>}}
\if{latex}{\out{\hypertarget{method-get_users}{}}}
\subsection{Method \code{get_users()}}{
Get information on users in the form of table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$get_users(users)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{users}}{A character vector of users}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
