% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineGraphQLGitLab.R
\name{EngineGraphQLGitLab}
\alias{EngineGraphQLGitLab}
\title{A EngineGraphQLGitLab class}
\description{
A class for methods wrapping GitLab's GraphQL API responses.
}
\section{Super class}{
\code{\link[GitStats:EngineGraphQL]{GitStats::EngineGraphQL}} -> \code{EngineGraphQLGitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineGraphQLGitLab$new()}}
\item \href{#method-get_orgs}{\code{EngineGraphQLGitLab$get_orgs()}}
\item \href{#method-get_repos}{\code{EngineGraphQLGitLab$get_repos()}}
\item \href{#method-get_repos_supportive}{\code{EngineGraphQLGitLab$get_repos_supportive()}}
\item \href{#method-get_commits}{\code{EngineGraphQLGitLab$get_commits()}}
\item \href{#method-clone}{\code{EngineGraphQLGitLab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineGraphQL" data-id="get_users">}\href{../../GitStats/html/EngineGraphQL.html#method-get_users}{\code{GitStats::EngineGraphQL$get_users()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineGraphQL" data-id="gql_response">}\href{../../GitStats/html/EngineGraphQL.html#method-gql_response}{\code{GitStats::EngineGraphQL$gql_response()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create \code{EngineGraphQLGitLab} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$new(gql_api_url, token, scan_all)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_api_url}}{GraphQL API url.}

\item{\code{token}}{A token.}

\item{\code{scan_all}}{A boolean.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_orgs"></a>}}
\if{latex}{\out{\hypertarget{method-get_orgs}{}}}
\subsection{Method \code{get_orgs()}}{
Get all groups from GitLab.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$get_orgs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to retrieve all repositories for an organization in
a table format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$get_repos(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_supportive"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_supportive}{}}}
\subsection{Method \code{get_repos_supportive()}}{
An empty method to satisfy engine iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$get_repos_supportive(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
Method to get commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$get_commits(
  org,
  date_from,
  date_until = Sys.date(),
  settings
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{date_from}}{A starting date to look commits for.}

\item{\code{date_until}}{An end date to look commits for.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method is empty as this class does not support pulling
commits - it is done for GitLab via REST. Still the method must
exist as it is called from the GitHost wrapper above.
}

\subsection{Returns}{
A table of commits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineGraphQLGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
