% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineRest.R
\name{EngineRest}
\alias{EngineRest}
\title{A EngineRest class}
\value{
A request.
}
\description{
A superclass for methods wraping Rest API responses.
}
\section{Super class}{
\code{\link[GitStats:Engine]{GitStats::Engine}} -> \code{EngineRest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gql_api_url}}{A character, url of Rest API.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{gql_api_url}}{A character, url of Rest API.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineRest$new()}}
\item \href{#method-response}{\code{EngineRest$response()}}
\item \href{#method-get_repos_by_phrase}{\code{EngineRest$get_repos_by_phrase()}}
\item \href{#method-find_by_id}{\code{EngineRest$find_by_id()}}
\item \href{#method-prepare_repos_table}{\code{EngineRest$prepare_repos_table()}}
\item \href{#method-clone}{\code{EngineRest$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Rest} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$new(rest_api_url = NA, token = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gql_api_url}}{A character, url of Rest API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Rest} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-response"></a>}}
\if{latex}{\out{\hypertarget{method-response}{}}}
\subsection{Method \code{response()}}{
A wrapper for httr2 functions to perform get request to REST API endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$response(endpoint, token = private$token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{An API endpoint.}

\item{\code{token}}{A token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A content of response formatted to list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_phrase"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_phrase}{}}}
\subsection{Method \code{get_repos_by_phrase()}}{
Method to get repositories with phrase in code blobs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$get_repos_by_phrase(phrase, org, language)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Table of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-find_by_id}{}}}
\subsection{Method \code{find_by_id()}}{
Perform get request to find projects by ids.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$find_by_id(ids, objects = c("repositories", "projects"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{A character vector of repositories or projects' ids.}

\item{\code{objects}}{A character to choose between 'repositories' (GitHub) and
'projects' (GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_repos_table"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_repos_table}{}}}
\subsection{Method \code{prepare_repos_table()}}{
A helper to prepare table for repositories content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$prepare_repos_table(repos_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_list}}{A repository list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
