% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineRest.R
\name{EngineRest}
\alias{EngineRest}
\title{A EngineRest class}
\value{
A request.

A request.
}
\description{
A superclass for methods wraping Rest API responses.
}
\section{Super class}{
\code{\link[GitStats:Engine]{GitStats::Engine}} -> \code{EngineRest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rest_api_url}}{A character, url of Rest API.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EngineRest$new()}}
\item \href{#method-response}{\code{EngineRest$response()}}
\item \href{#method-get_repos}{\code{EngineRest$get_repos()}}
\item \href{#method-get_commits}{\code{EngineRest$get_commits()}}
\item \href{#method-clone}{\code{EngineRest$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Rest} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$new(rest_api_url = NA, token = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rest_api_url}}{A character, url of Rest API.}

\item{\code{token}}{A token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Rest} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-response"></a>}}
\if{latex}{\out{\hypertarget{method-response}{}}}
\subsection{Method \code{response()}}{
A wrapper for httr2 functions to perform get request to REST API endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$response(endpoint, token = private$token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{An API endpoint.}

\item{\code{token}}{A token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A content of response formatted to list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
Method to get repositories with phrase in code blobs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$get_repos(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Table of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$get_commits(
  org,
  repos_table,
  date_from,
  date_until = Sys.date(),
  settings
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
