% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineRestGitHub.R
\name{EngineRestGitHub}
\alias{EngineRestGitHub}
\title{A EngineRestGitHub class}
\description{
A class for methods wraping GitHub's REST API responses.
}
\section{Super classes}{
\code{\link[GitStats:Engine]{GitStats::Engine}} -> \code{\link[GitStats:EngineRest]{GitStats::EngineRest}} -> \code{EngineRestGitHub}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_repos_contributors}{\code{EngineRestGitHub$get_repos_contributors()}}
\item \href{#method-get_repos_issues}{\code{EngineRestGitHub$get_repos_issues()}}
\item \href{#method-clone}{\code{EngineRestGitHub$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="get_repos_by_phrase">}\href{../../GitStats/html/EngineRest.html#method-get_repos_by_phrase}{\code{GitStats::EngineRest$get_repos_by_phrase()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="initialize">}\href{../../GitStats/html/EngineRest.html#method-initialize}{\code{GitStats::EngineRest$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="response">}\href{../../GitStats/html/EngineRest.html#method-response}{\code{GitStats::EngineRest$response()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_contributors}{}}}
\subsection{Method \code{get_repos_contributors()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitHub$get_repos_contributors(repos_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_table}}{A table of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_issues"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_issues}{}}}
\subsection{Method \code{get_repos_issues()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitHub$get_repos_issues(repos_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_table}}{A table of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitHub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
