% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineRestGitLab.R
\name{EngineRestGitLab}
\alias{EngineRestGitLab}
\title{A EngineRestGitLab class}
\description{
A class for methods wraping GitLab's REST API responses.
}
\section{Super classes}{
\code{\link[GitStats:Engine]{GitStats::Engine}} -> \code{\link[GitStats:EngineRest]{GitStats::EngineRest}} -> \code{EngineRestGitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_repos_by_phrase}{\code{EngineRestGitLab$get_repos_by_phrase()}}
\item \href{#method-get_repos_contributors}{\code{EngineRestGitLab$get_repos_contributors()}}
\item \href{#method-get_repos_issues}{\code{EngineRestGitLab$get_repos_issues()}}
\item \href{#method-get_commits_from_org}{\code{EngineRestGitLab$get_commits_from_org()}}
\item \href{#method-clone}{\code{EngineRestGitLab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="initialize">}\href{../../GitStats/html/EngineRest.html#method-initialize}{\code{GitStats::EngineRest$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="response">}\href{../../GitStats/html/EngineRest.html#method-response}{\code{GitStats::EngineRest$response()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_phrase"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_phrase}{}}}
\subsection{Method \code{get_repos_by_phrase()}}{
Method to get repositories with phrase in code blobs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$get_repos_by_phrase(phrase, org)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phrase}}{A phrase to look for in codelines.}

\item{\code{org}}{A character, an organization of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
For the time being there is no possibility to search GitLab with
filtering by language. For more information look here:
\link{https://gitlab.com/gitlab-org/gitlab/-/issues/340333}
}

\subsection{Returns}{
Table of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_contributors}{}}}
\subsection{Method \code{get_repos_contributors()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$get_repos_contributors(repos_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_table}}{A table of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_issues"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_issues}{}}}
\subsection{Method \code{get_repos_issues()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$get_repos_issues(repos_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_table}}{A table of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits_from_org"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits_from_org}{}}}
\subsection{Method \code{get_commits_from_org()}}{
GitLab private method to derive
commits from repo with REST API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$get_commits_from_org(
  org,
  repos_table,
  date_from,
  date_until = Sys.date(),
  by,
  team
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A character, a group of projects.}

\item{\code{repos_table}}{A table of repositories.}

\item{\code{date_from}}{A starting date to look commits for.}

\item{\code{date_until}}{An end date to look commits for.}

\item{\code{by}}{A search parameter - \code{team} or \code{orgs}.}

\item{\code{team}}{A list of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of commits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
