% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineRestGitLab.R
\name{EngineRestGitLab}
\alias{EngineRestGitLab}
\title{A EngineRestGitLab class}
\description{
A class for methods wrapping GitLab's REST API responses.
}
\section{Super class}{
\code{\link[GitStats:EngineRest]{GitStats::EngineRest}} -> \code{EngineRestGitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pull_repos}{\code{EngineRestGitLab$pull_repos()}}
\item \href{#method-pull_repos_supportive}{\code{EngineRestGitLab$pull_repos_supportive()}}
\item \href{#method-pull_repos_contributors}{\code{EngineRestGitLab$pull_repos_contributors()}}
\item \href{#method-pull_commits}{\code{EngineRestGitLab$pull_commits()}}
\item \href{#method-clone}{\code{EngineRestGitLab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="check_organizations">}\href{../../GitStats/html/EngineRest.html#method-check_organizations}{\code{GitStats::EngineRest$check_organizations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="initialize">}\href{../../GitStats/html/EngineRest.html#method-initialize}{\code{GitStats::EngineRest$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="EngineRest" data-id="response">}\href{../../GitStats/html/EngineRest.html#method-response}{\code{GitStats::EngineRest$response()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_repos"></a>}}
\if{latex}{\out{\hypertarget{method-pull_repos}{}}}
\subsection{Method \code{pull_repos()}}{
A method to retrieve all repositories for an organization in
a table format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$pull_repos(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A character, a group of projects.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_repos_supportive"></a>}}
\if{latex}{\out{\hypertarget{method-pull_repos_supportive}{}}}
\subsection{Method \code{pull_repos_supportive()}}{
An empty method to satisfy engine iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$pull_repos_supportive(org, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_repos_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-pull_repos_contributors}{}}}
\subsection{Method \code{pull_repos_contributors()}}{
A method to add information on repository contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$pull_repos_contributors(repos_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_table}}{A table of repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of repositories with added information on contributors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_commits"></a>}}
\if{latex}{\out{\hypertarget{method-pull_commits}{}}}
\subsection{Method \code{pull_commits()}}{
GitLab private method to derive
commits from repo with REST API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$pull_commits(
  org,
  date_from,
  date_until = Sys.date(),
  settings
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A character, a group of projects.}

\item{\code{date_from}}{A starting date to look commits for.}

\item{\code{date_until}}{An end date to look commits for.}

\item{\code{settings}}{A list of  \code{GitStats} settings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table of commits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineRestGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
