% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_commits}
\alias{get_commits}
\title{Get commits}
\usage{
get_commits(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of commits.
}
\description{
Retrieves commits table pulled by \code{GitStats}.
}
