% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commits.R
\name{get_commits}
\alias{get_commits}
\title{Get information on commits.}
\usage{
get_commits(
  gitstats_obj,
  date_from = NULL,
  date_until = Sys.time(),
  by = gitstats_obj$search_param,
  print_out = TRUE
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{date_from}{A starting date to look commits for}

\item{date_until}{An end date to look commits for}

\item{by}{A character, to choose between: \itemize{\item{org - organizations
(owners of repositories)} \item{team - A team} \item{phrase - A keyword in
code blobs.}}}

\item{print_out}{A boolean stating whether to print an output.}
}
\value{
A \code{GitStats} class object with updated \verb{$commits_dt} field.
}
\description{
List all repositories for an organization, a team.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_connection(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "erasmusmc-public-health"
  ) \%>\%
  get_commits(date_from = "2020-01-01")
}
}
