% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_commits_stats}
\alias{get_commits_stats}
\title{Get statistics on commits}
\usage{
get_commits_stats(gitstats_obj, time_interval = c("month", "day", "week"))
}
\arguments{
\item{gitstats_obj}{A GitStats class object.}

\item{time_interval}{A character, specifying time interval to show statistics.}
}
\value{
A table of \code{commits_stats} class.
}
\description{
Prepare statistics from the pulled commits data.
}
