% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_commits_stats}
\alias{get_commits_stats}
\title{Get commits statistics}
\usage{
get_commits_stats(
  gitstats_object,
  time_aggregation = c("year", "month", "day", "week"),
  ...,
  stats = dplyr::n()
)
}
\arguments{
\item{gitstats_object}{A \code{GitStats} object.}

\item{time_aggregation}{A character, specifying time aggregation of
statistics.}

\item{...}{Other grouping variables to be passed to \code{dplyr::group_by()}
function apart from \code{stats_date} and \code{githost}. Could be: \code{author},
\code{author_login}, \code{author_name}, \code{repository} or \code{organization}. Should be
passed without quotation marks.}

\item{stats}{Customize statistics.}
}
\value{
A table of \code{commits_stats} class.
}
\description{
Prepare statistics from the pulled commits data.
}
\details{
To make function work, you need first to get commits data with
\code{GitStats}. See examples section.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
   set_github_host(
     token = Sys.getenv("GITHUB_PAT"),
     repos = c("r-world-devs/GitStats", "openpharma/visR")
   )
 get_commits(my_gitstats, since = "2022-01-01")
 get_commits_stats(
   gitstats_object = my_gitstats,
   author,
   organization,
   time_interval = "year"
 )
}
}
