% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Get files}
\usage{
get_files(
  gitstats,
  pattern = NULL,
  depth = Inf,
  file_path = NULL,
  cache = TRUE,
  verbose = is_verbose(gitstats),
  progress = verbose
)
}
\arguments{
\item{gitstats}{A \code{GitStats} object.}

\item{pattern}{A regular expression. If defined, it pulls content of all
files in a repository matching this pattern reaching to the level of
directories defined by \code{depth} parameter. Works only if \code{file_path} stays
\code{NULL}.}

\item{depth}{Defines level of directories to retrieve files from. E.g. if set
to \code{0}, it will pull files only from root, if \code{1L}, will take data from
\code{root} directory and directories visible in \code{root} directory. If left with
no argument, will pull files from all directories.}

\item{file_path}{A specific path to file(s) in repositories. If defined, the
function pulls content of this specific \code{file_path}. May be a character
vector if multiple files are to be pulled. Can be defined only if \code{pattern}
stays \code{NULL}.}

\item{cache}{A logical, if set to \code{TRUE} GitStats will retrieve the last
result from its storage.}

\item{verbose}{A logical, \code{TRUE} by default. If \code{FALSE} messages and printing
output is switched off.}

\item{progress}{A logical, by default set to \code{verbose} value. If \code{FALSE} no
\code{cli} progress bar will be displayed.}
}
\value{
A data.frame.
}
\description{
Pulls text files and their content.
}
\examples{
\dontrun{
 git_stats <- create_gitstats() |>
  set_github_host(
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) |>
  set_gitlab_host(
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  )

 rmd_files <- get_files(
   gitstats = git_stats,
   pattern = "\\\\.Rmd",
   depth = 2L
 )

 app_files <- get_files(
   gitstats = git_stats,
   file_path = "R/app.R"
 )

}
}
