% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_orgs}
\alias{get_orgs}
\title{Get organizations}
\usage{
get_orgs(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A vector of organizations.
}
\description{
Retrieves organizations set or pulled by \code{GitStats}. Especially
helpful when user is scanning whole git platform and wants to have a
glimpse at organizations.
}
