% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_R_package_usage}
\alias{get_R_package_usage}
\title{Get R package usage}
\usage{
get_R_package_usage(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table with repo urls.
}
\description{
Retrieves list of repositories that make use of a package.
}
