% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_release_logs}
\alias{get_release_logs}
\title{Get release logs}
\usage{
get_release_logs(
  gitstats_object,
  since = NULL,
  until = NULL,
  use_storage = TRUE,
  verbose = is_verbose(gitstats_object)
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{since}{A starting date.}

\item{until}{An end date.}

\item{use_storage}{A logical, should storage be used.}

\item{verbose}{A logical, should verbose mode be used.}
}
\value{
A table with release logs.
}
\description{
Pull release logs from repositories.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  )
  get_release_logs(my_gistats, since = "2024-01-01")
}
}
