% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_release_logs}
\alias{get_release_logs}
\title{Get release logs}
\usage{
get_release_logs(
  gitstats_object,
  since = NULL,
  until = NULL,
  cache = TRUE,
  verbose = is_verbose(gitstats_object)
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{since}{A starting date.}

\item{until}{An end date.}

\item{cache}{A logical, if \code{TRUE}, cached results will be used.}

\item{verbose}{A logical, \code{TRUE} by default. If \code{FALSE} messages and
printing output is switched off.}
}
\value{
A table with release logs.
}
\description{
Pull release logs from repositories.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_github_host(
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  )
  get_release_logs(my_gistats, since = "2024-01-01")
}
}
