% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_repos}
\alias{get_repos}
\title{Get repositories}
\usage{
get_repos(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of repositories.
}
\description{
Retrieves repositories table pulled by \code{GitStats}.
}
\examples{
create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_repos() \%>\%
  get_repos()
}
