% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{get_repos}
\alias{get_repos}
\title{Get information on repositories}
\usage{
get_repos(
  gitstats_obj,
  by = gitstats_obj$search_param,
  phrase = gitstats_obj$phrase,
  language = gitstats_obj$language,
  print_out = TRUE
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{by}{A character, to choose between: \itemize{\item{org - organizations
(owners of repositories)} \item{team - A team} \item{phrase - A keyword in
code blobs.}}}

\item{phrase}{A character, phrase to look for in codelines.}

\item{language}{A character specifying language used in repositories.}

\item{print_out}{A boolean stating whether to print an output.}
}
\value{
A \code{GitStats} class object with updated \verb{$repos_dt} field.
}
\description{
List all repositories for an organization, a team or by a
keyword.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_connection(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT"),
    orgs = "erasmusmc-public-health"
  ) \%>\%
  get_repos(language = "R")

my_gitstats$get_repos(
  by = "phrase",
  phrase = "diabetes",
  language = "R"
)
}
}
