% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_repos}
\alias{get_repos}
\title{Get information on repositories}
\usage{
get_repos(
  gitstats_object,
  add_contributors = TRUE,
  with_code = NULL,
  use_storage = TRUE,
  verbose = is_verbose(gitstats_object)
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{add_contributors}{A logical parameter to decide whether to add
information about repositories' contributors to the repositories output
(table). If set to \code{FALSE} it makes function run faster as, in the case of
\code{org} search mode, it reaches only \code{GraphQL} endpoint with a query on
repositories, and in the case of \code{code} search mode it reaches only
\verb{repositories REST API} endpoint. However, the pitfall is that the result
does not convey information on contributors. \cr\cr When set to \code{TRUE} (by
default), \code{GitStats} iterates additionally over pulled repositories and
reaches to the \verb{contributors APIs}, which makes it slower, but gives
additional information.}

\item{with_code}{A character, if  defined, GitStats will pull repositories
with specified text in code blobs.}

\item{use_storage}{A logical, should storage be used.}

\item{verbose}{A logical, should verbose mode be used.}
}
\description{
List all repositories for an organization or by a keyword.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  )
  get_repos(my_gitstats)
}
}
