% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_repos_stats}
\alias{get_repos_stats}
\title{Get Statistics On Repositories}
\usage{
get_repos_stats(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats class object.}
}
\value{
A table of \code{repos_stats} class.
}
\description{
Prepare statistics from the pulled repositories data.
}
