% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_repos_urls}
\alias{get_repos_urls}
\title{List repository URLS}
\usage{
get_repos_urls(gitstats_object, with_file = NULL)
}
\arguments{
\item{with_file}{A character, if  defined, GitStats will pull
repositories with specified file.}
}
\value{
A character vector.
}
\description{
A wrapper over search API endpoints to list repositories
URLS with a given file.
}
