% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_repos_urls}
\alias{get_repos_urls}
\title{List repository URLS}
\usage{
get_repos_urls(
  gitstats_object,
  type = "web",
  with_files = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gitstats_object}{A GitStats object.}

\item{type}{A character, choose if \code{api} or \code{web} (\code{html}) URLs should be
returned.}

\item{with_files}{A character vector, if defined, GitStats will pull
repositories with specified files.}

\item{verbose}{A logical, \code{TRUE} by default. If \code{FALSE} messages and printing
output is switched off.}
}
\value{
A character vector.
}
\description{
A wrapper over search API endpoints to list repositories URLS.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_github_host(
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  set_gitlab_host(
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  )
get_repos_urls(my_gitstats, type = "api")
get_repos_urls(my_gitstats, with_files = c("DESCRIPTION", "LICENSE"))
}
}
