% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_users}
\alias{get_users}
\title{Get users}
\usage{
get_users(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A table of users.
}
\description{
Retrieves users table pulled by \code{GitStats}.
}
