% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{get_users}
\alias{get_users}
\title{Get users statistics.}
\usage{
get_users(gitstats_obj, users)
}
\arguments{
\item{users}{A character vector of users.}
}
\value{
A table of users.
}
\description{
Get information on users.
}
\examples{
\dontrun{
 my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_connection(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  get_users(c("maciekabanas", "marcinkowskak"))
}
}
