% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub.R
\name{GitHub}
\alias{GitHub}
\title{A GitHub API Client class}
\description{
An object with methods to obtain information form GitHub API.
}
\section{Super class}{
\code{\link[GitStats:GitPlatform]{GitStats::GitPlatform}} -> \code{GitHub}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitHub$new()}}
\item \href{#method-get_repos}{\code{GitHub$get_repos()}}
\item \href{#method-get_commits}{\code{GitHub$get_commits()}}
\item \href{#method-print}{\code{GitHub$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitHub} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHub$new(rest_api_url = NA, gql_api_url = NA, token = NA, orgs = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rest_api_url}}{A url of rest API.}

\item{\code{gql_api_url}}{A url of GraphQL API.}

\item{\code{token}}{A token.}

\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GitHub} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to list all repositories for an organization, a
team or by a keyword.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHub$get_repos(by, team, phrase, language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{A character, to choose between: \itemize{\item{org -
organizations (project groups)} \item{team -
A team} \item{phrase - A keyword in code blobs.}}}

\item{\code{team}}{A list of team members.}

\item{\code{phrase}}{A character to look for in code blobs. Obligatory if
\code{by} parameter set to \code{"phrase"}.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHub$get_commits(date_from, date_until = Sys.Date(), by, team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{A starting date to look commits for.}

\item{\code{date_until}}{An end date to look commits for.}

\item{\code{by}}{A character, to choose between: \itemize{\item{org -
organizations (owners of repositories or project groups)} \item{team -
A team} \item{phrase - A keyword in code blobs.}}}

\item{\code{team}}{A list of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of commits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitHub object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitHub$print()}\if{html}{\out{</div>}}
}

}
}
