% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitLab.R
\name{GitLab}
\alias{GitLab}
\title{A GitLab API Client class}
\description{
An object with methods to obtain information form GitLab API.
}
\section{Super class}{
\code{\link[GitStats:GitService]{GitStats::GitService}} -> \code{GitLab}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{repos_endpoint}}{An expression for repositories endpoint.}

\item{\code{repo_contributors_endpoint}}{An expression for repositories contributors endpoint.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{GitLab$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitService" data-id="get_commits">}\href{../../GitStats/html/GitService.html#method-get_commits}{\code{GitStats::GitService$get_commits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitService" data-id="get_repos">}\href{../../GitStats/html/GitService.html#method-get_repos}{\code{GitStats::GitService$get_repos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitService" data-id="initialize">}\href{../../GitStats/html/GitService.html#method-initialize}{\code{GitStats::GitService$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitLab object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$print()}\if{html}{\out{</div>}}
}

}
}
