% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitLab.R
\name{GitLab}
\alias{GitLab}
\title{A GitLab API Client class}
\description{
An object with methods to obtain information form GitLab API.
}
\section{Super class}{
\code{\link[GitStats:GitService]{GitStats::GitService}} -> \code{GitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_repos}{\code{GitLab$get_repos()}}
\item \href{#method-get_commits}{\code{GitLab$get_commits()}}
\item \href{#method-print}{\code{GitLab$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GitStats" data-topic="GitService" data-id="initialize">}\href{../../GitStats/html/GitService.html#method-initialize}{\code{GitStats::GitService$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to list all repositories for an organization, a
team or by a keyword.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_repos(orgs = self$orgs, by, team, phrase, language = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orgs}}{A character vector of organizations (owners of repositories).}

\item{\code{by}}{A character, to choose between: \itemize{\item{org -
organizations (owners of repositories or project groups)} \item{team -
A team} \item{phrase - A keyword in code blobs.}}}

\item{\code{team}}{A list of team members.}

\item{\code{phrase}}{A character to look for in code blobs. Obligatory if
\code{by} parameter set to \code{"phrase"}.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$get_commits(
  orgs = self$orgs,
  date_from,
  date_until = Sys.time(),
  by,
  team
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orgs}}{A character vector of organisations.}

\item{\code{date_from}}{A starting date to look commits for}

\item{\code{date_until}}{An end date to look commits for}

\item{\code{by}}{A character, to choose between: \itemize{\item{org -
organizations (owners of repositories or project groups)} \item{team -
A team} \item{phrase - A keyword in code blobs.}}}

\item{\code{team}}{A list of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of commits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitLab object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitLab$print()}\if{html}{\out{</div>}}
}

}
}
