% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{GitStats}
\alias{GitStats}
\title{A statistics platform for Git hosts}
\description{
An R6 class object with methods to derive information from multiple Git platforms.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setup}{\code{GitStats$setup()}}
\item \href{#method-add_host}{\code{GitStats$add_host()}}
\item \href{#method-add_team_member}{\code{GitStats$add_team_member()}}
\item \href{#method-get_repos}{\code{GitStats$get_repos()}}
\item \href{#method-get_commits}{\code{GitStats$get_commits()}}
\item \href{#method-show_repos}{\code{GitStats$show_repos()}}
\item \href{#method-show_commits}{\code{GitStats$show_commits()}}
\item \href{#method-print}{\code{GitStats$print()}}
\item \href{#method-clone}{\code{GitStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Set up your search settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$setup(search_param, team_name, phrase, language, print_out)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_param}}{One of three: \code{team}, \code{org} or \code{phrase}.}

\item{\code{team_name}}{A name of a team.}

\item{\code{phrase}}{A phrase to look for.}

\item{\code{language}}{A language of programming code.}

\item{\code{print_out}}{A boolean stating if you want to print output after
pulling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_host"></a>}}
\if{latex}{\out{\hypertarget{method-add_host}{}}}
\subsection{Method \code{add_host()}}{
Method to set connections to Git platforms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$add_host(api_url, token, orgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_url}}{A character, url address of API.}

\item{\code{token}}{A token.}

\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, puts connection information into \verb{$hosts} slot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_team_member"></a>}}
\if{latex}{\out{\hypertarget{method-add_team_member}{}}}
\subsection{Method \code{add_team_member()}}{
A method to add a team member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$add_team_member(member_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{member_name}}{Name of a member.}

\item{\code{...}}{User names on Git platforms.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, pass information on team member to \code{GitStats}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to list all repositories for an organization,
a team or by a keyword.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_commits(date_from, date_until)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{A starting date to look commits for}

\item{\code{date_until}}{An end date to look commits for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of commits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_repos"></a>}}
\if{latex}{\out{\hypertarget{method-show_repos}{}}}
\subsection{Method \code{show_repos()}}{
Print repositories output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$show_repos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_commits"></a>}}
\if{latex}{\out{\hypertarget{method-show_commits}{}}}
\subsection{Method \code{show_commits()}}{
Print commits output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$show_commits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitStats object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
