% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gitstats.R
\name{GitStats}
\alias{GitStats}
\title{A statistics platform for Git clients}
\description{
An R6 class object with methods to derive information from multiple Git platforms.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GitStats-new}{\code{GitStats$new()}}
\item \href{#method-GitStats-get_repos_by_owner_or_group}{\code{GitStats$get_repos_by_owner_or_group()}}
\item \href{#method-GitStats-get_repos_by_team}{\code{GitStats$get_repos_by_team()}}
\item \href{#method-GitStats-get_repos_by_codephrase}{\code{GitStats$get_repos_by_codephrase()}}
\item \href{#method-GitStats-get_commits_by_owner_or_group}{\code{GitStats$get_commits_by_owner_or_group()}}
\item \href{#method-GitStats-get_commits_by_team}{\code{GitStats$get_commits_by_team()}}
\item \href{#method-GitStats-set_team}{\code{GitStats$set_team()}}
\item \href{#method-GitStats-plot_repos}{\code{GitStats$plot_repos()}}
\item \href{#method-GitStats-plot_commits}{\code{GitStats$plot_commits()}}
\item \href{#method-GitStats-plot_commit_lines}{\code{GitStats$plot_commit_lines()}}
\item \href{#method-GitStats-print}{\code{GitStats$print()}}
\item \href{#method-GitStats-clone}{\code{GitStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-new"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitStats} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A list of objects of class GitHubClient or GitLabClient}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GitStats} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-get_repos_by_owner_or_group"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-get_repos_by_owner_or_group}{}}}
\subsection{Method \code{get_repos_by_owner_or_group()}}{
A method to list all repositories for an organization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos_by_owner_or_group()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-get_repos_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-get_repos_by_team}{}}}
\subsection{Method \code{get_repos_by_team()}}{
A method to list all repositories
for a team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos_by_team(team_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_name}}{A name of a team}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-get_repos_by_codephrase"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-get_repos_by_codephrase}{}}}
\subsection{Method \code{get_repos_by_codephrase()}}{
A method to find repositories with given phrase in codelines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos_by_codephrase(phrase, language = "R")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phrase}}{A phrase to look for in codelines.}

\item{\code{language}}{A character specifying language used in repositories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-get_commits_by_owner_or_group"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-get_commits_by_owner_or_group}{}}}
\subsection{Method \code{get_commits_by_owner_or_group()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_commits_by_owner_or_group(date_from, date_until = Sys.time())}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-get_commits_by_team"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-get_commits_by_team}{}}}
\subsection{Method \code{get_commits_by_team()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_commits_by_team(team_name, date_from, date_until = Sys.time())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-set_team"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-set_team}{}}}
\subsection{Method \code{set_team()}}{
A method to set your team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_team(team_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_name}}{A name of a team.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-plot_repos"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-plot_repos}{}}}
\subsection{Method \code{plot_repos()}}{
A method to plot repositories outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_repos(repos_dt = self$repos_dt, repos_n = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_dt}}{An output of one of \code{get_repos} methods.}

\item{\code{repos_n}}{An integer, a number of repos to show on the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-plot_commits"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-plot_commits}{}}}
\subsection{Method \code{plot_commits()}}{
A method to plot basic commit stats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_commits(
  commits_dt = self$commits_dt,
  stats_by = c("day", "week", "month")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-plot_commit_lines"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-plot_commit_lines}{}}}
\subsection{Method \code{plot_commit_lines()}}{
A method to plot commit additions and deletions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_commit_lines(commits_dt = self$commits_dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commits_dt}}{An output of one of \code{get_commits} methods.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-print"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitStats object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GitStats-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GitStats-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
