% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gitstats.R
\name{GitStats}
\alias{GitStats}
\title{A statistics platform for Git clients}
\description{
An R6 class object with methods to derive information from multiple Git platforms.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitStats$new()}}
\item \href{#method-get_repos}{\code{GitStats$get_repos()}}
\item \href{#method-get_repos_by_codephrase}{\code{GitStats$get_repos_by_codephrase()}}
\item \href{#method-get_commits}{\code{GitStats$get_commits()}}
\item \href{#method-set_connection}{\code{GitStats$set_connection()}}
\item \href{#method-set_team}{\code{GitStats$set_team()}}
\item \href{#method-plot_repos}{\code{GitStats$plot_repos()}}
\item \href{#method-plot_commits}{\code{GitStats$plot_commits()}}
\item \href{#method-plot_commit_lines}{\code{GitStats$plot_commit_lines()}}
\item \href{#method-print}{\code{GitStats$print()}}
\item \href{#method-clone}{\code{GitStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GitStats} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{GitStats} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos}{}}}
\subsection{Method \code{get_repos()}}{
A method to list all repositories by an organization or a
team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos(by = "org", print_out = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{A chararacter, to choose between: \itemize{
\item{org}{Organizations} \item(team){Team}}}

\item{\code{print_out}}{A boolean stating whether to print an output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repos_by_codephrase"></a>}}
\if{latex}{\out{\hypertarget{method-get_repos_by_codephrase}{}}}
\subsection{Method \code{get_repos_by_codephrase()}}{
A method to find repositories with given phrase in codelines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_repos_by_codephrase(phrase, language = "R", print_out = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phrase}}{A phrase to look for in codelines.}

\item{\code{language}}{A character specifying language used in repositories.}

\item{\code{print_out}}{A boolean stating whether to print an output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of repositories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_commits"></a>}}
\if{latex}{\out{\hypertarget{method-get_commits}{}}}
\subsection{Method \code{get_commits()}}{
A method to get information on commits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$get_commits(
  date_from = NULL,
  date_until = Sys.time(),
  by = "org",
  print_out = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{A starting date to look commits for}

\item{\code{date_until}}{An end date to look commits for}

\item{\code{by}}{A chararacter, to choose between: \itemize{
\item{org}{Organizations} \item(team){Team}}}

\item{\code{print_out}}{A boolean stating whether to print an output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of commits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_connection"></a>}}
\if{latex}{\out{\hypertarget{method-set_connection}{}}}
\subsection{Method \code{set_connection()}}{
Method to set connections to Git platforms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_connection(api_url, token, orgs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_url}}{A character, url address of API.}

\item{\code{token}}{A token.}

\item{\code{orgs}}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, puts connection information into \verb{$clients} slot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_team"></a>}}
\if{latex}{\out{\hypertarget{method-set_team}{}}}
\subsection{Method \code{set_team()}}{
A method to set your team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$set_team(team_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_name}}{A name of a team.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, puts team information into \verb{$team} slot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_repos"></a>}}
\if{latex}{\out{\hypertarget{method-plot_repos}{}}}
\subsection{Method \code{plot_repos()}}{
A method to plot repositories outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_repos(repos_dt = self$repos_dt, repos_n = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos_dt}}{An output of one of \code{get_repos} methods.}

\item{\code{repos_n}}{An integer, a number of repos to show on the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_commits"></a>}}
\if{latex}{\out{\hypertarget{method-plot_commits}{}}}
\subsection{Method \code{plot_commits()}}{
A method to plot basic commit stats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_commits(
  commits_dt = self$commits_dt,
  stats_by = c("day", "week", "month")
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_commit_lines"></a>}}
\if{latex}{\out{\hypertarget{method-plot_commit_lines}{}}}
\subsection{Method \code{plot_commit_lines()}}{
A method to plot commit additions and deletions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$plot_commit_lines(commits_dt = self$commits_dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commits_dt}}{An output of one of \code{get_commits} methods.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
A print method for a GitStats object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
