% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats.R
\name{gitstats}
\alias{gitstats}
\title{Plot github stats}
\usage{
gitstats(id = "hadley", type = c("users", "orgs"), max = 100, ...)
}
\arguments{
\item{id}{name of the github user or organization. Default: "Hadley"}

\item{type}{either "users" (default) or "orgs"}

\item{max}{maximum number of repositories to plot. Default: 100. Max: 100}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Plots the total number of forks and stars for all repositories of 
a user or organization.
}
\examples{
\dontrun{
gitstats(max = 10)
gitstats(max = 40)
gitstats(id = "jeroen", max = 30)
gitstats(id = "ropensci", type = "orgs", max = 70)
}
}
\author{
Scott Chamberlain, Jeroen Ooms
}
