% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gitstats_plot}
\alias{gitstats_plot}
\title{Plot Git statistics}
\usage{
gitstats_plot(
  stats_table = NULL,
  value_to_plot = NULL,
  plotly_mode = FALSE,
  value_decreasing = TRUE,
  n = NULL
)
}
\arguments{
\item{stats_table}{A table with repository or commits statistics.}

\item{value_to_plot}{Value to be plotted.}

\item{plotly_mode}{A boolean, if TRUE, turns plot into interactive plotly.}

\item{value_decreasing}{A boolean to set ordering of a value on the plot.}

\item{n}{An integer, a maximum number of repos/organizations to show on the plot.}
}
\value{
A plot.
}
\description{
A generic to plot statistics from repositories or commits.
}
\examples{
 repos_stats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_repos() \%>\%
  get_repos_stats()

  gitstats_plot(
    stats_table = repos_stats,
    value_to_plot = "contributors_n",
    value_decreasing = FALSE
  )
}
