% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GQLQueryGitLab.R
\name{GQLQueryGitLab}
\alias{GQLQueryGitLab}
\title{A GQLQueryGitLab class}
\description{
A class with methods to build GraphQL Queries for GitLab.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-repos_by_org}{\code{GQLQueryGitLab$repos_by_org()}}
\item \href{#method-user}{\code{GQLQueryGitLab$user()}}
\item \href{#method-clone}{\code{GQLQueryGitLab$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repos_by_org"></a>}}
\if{latex}{\out{\hypertarget{method-repos_by_org}{}}}
\subsection{Method \code{repos_by_org()}}{
Prepare query to get repositories from GitLab.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitLab$repos_by_org(repo_cursor = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo_cursor}}{An end cursor for repositories page.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user"></a>}}
\if{latex}{\out{\hypertarget{method-user}{}}}
\subsection{Method \code{user()}}{
Prepare query to get info on a GitLab user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitLab$user()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
