% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GQLQueryGitLab.R
\name{GQLQueryGitLab}
\alias{GQLQueryGitLab}
\title{A GQLQueryGitLab class}
\description{
A class with methods to build GraphQL Queries for GitLab.
}
\section{Super class}{
\code{\link[GitStats:GQLQuery]{GitStats::GQLQuery}} -> \code{GQLQueryGitLab}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-projects_by_group}{\code{GQLQueryGitLab$projects_by_group()}}
\item \href{#method-clone}{\code{GQLQueryGitLab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projects_by_group"></a>}}
\if{latex}{\out{\hypertarget{method-projects_by_group}{}}}
\subsection{Method \code{projects_by_group()}}{
Method to build query to pull projects by group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitLab$projects_by_group(group, projects_cursor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{A group of projects.}

\item{\code{projects_cursor}}{A cursor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GQLQueryGitLab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
