% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQLQuery.R
\name{GraphQLQuery}
\alias{GraphQLQuery}
\title{A GraphQL class}
\description{
An object with methods to build GraphQL Queries.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-repos_by_org}{\code{GraphQLQuery$repos_by_org()}}
\item \href{#method-users_id}{\code{GraphQLQuery$users_id()}}
\item \href{#method-commits_by_repo}{\code{GraphQLQuery$commits_by_repo()}}
\item \href{#method-groups_by_user}{\code{GraphQLQuery$groups_by_user()}}
\item \href{#method-projects_by_group}{\code{GraphQLQuery$projects_by_group()}}
\item \href{#method-clone}{\code{GraphQLQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repos_by_org"></a>}}
\if{latex}{\out{\hypertarget{method-repos_by_org}{}}}
\subsection{Method \code{repos_by_org()}}{
Prepare query to pull repositories for GitHub organization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$repos_by_org(org, language, repo_cursor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{An organization.}

\item{\code{language}}{Language of a repository.}

\item{\code{repo_cursor}}{An endCursor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-users_id"></a>}}
\if{latex}{\out{\hypertarget{method-users_id}{}}}
\subsection{Method \code{users_id()}}{
Prepare query to get ID of a GitHub user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$users_id(login)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{login}}{A login of a user.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commits_by_repo"></a>}}
\if{latex}{\out{\hypertarget{method-commits_by_repo}{}}}
\subsection{Method \code{commits_by_repo()}}{
Prepare query to get commits on GitHub
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$commits_by_repo(
  org,
  repo,
  since,
  until,
  commits_cursor = "",
  author_id = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{A GitHub organization}

\item{\code{repo}}{Name of a repository}

\item{\code{since}}{Git Time Stamp of starting date of commits.}

\item{\code{until}}{Git Time STamp of end date of commits.}

\item{\code{commits_cursor}}{An endCursor.}

\item{\code{author_id}}{An Id of an author.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_by_user"></a>}}
\if{latex}{\out{\hypertarget{method-groups_by_user}{}}}
\subsection{Method \code{groups_by_user()}}{
GitLab. Method to build query to pull groups by users.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$groups_by_user(team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A string of team members.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projects_by_group"></a>}}
\if{latex}{\out{\hypertarget{method-projects_by_group}{}}}
\subsection{Method \code{projects_by_group()}}{
GitLab. Method to build query to pull projects by group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$projects_by_group(group, projects_cursor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{A group of projects.}

\item{\code{projects_cursor}}{A cursor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphQLQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
