% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commits.R
\name{plot_commits}
\alias{plot_commits}
\title{Plot commits data.}
\usage{
plot_commits(gitstats_obj, stats_by = c("day", "week", "month"))
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{stats_by}{A character, specifing time interval to show statistics.}
}
\value{
A plot.
}
\description{
Plot commits data.
}
