% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_commits_stats}
\alias{plot_commits_stats}
\title{Plot commits statistics}
\usage{
plot_commits_stats(
  stats_table = NULL,
  value_to_plot = "commits_n",
  plotly_mode = FALSE
)
}
\arguments{
\item{stats_table}{A table with commits statistics.}

\item{value_to_plot}{Value to be plotted.}

\item{plotly_mode}{A boolean, if TRUE, turns plot into interactive plotly.}
}
\value{
A plot.
}
\description{
A function to plot statistics from commits.
}
