% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_repos}
\alias{plot_repos}
\title{Plot repository data.}
\usage{
plot_repos(gitstats_obj, repos_n = 10)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{repos_n}{An integer, a number of repos to show on the plot.}
}
\value{
A plot.
}
\description{
Plots top repositories by last activity.
}
