% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_commits}
\alias{pull_commits}
\title{Get information on commits}
\usage{
pull_commits(gitstats_obj, date_from = NULL, date_until = Sys.time())
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{date_from}{A starting date to look commits for}

\item{date_until}{An end date to look commits for}
}
\value{
A \code{GitStats} class object with updated \verb{$commits} field.
}
\description{
List all commits from all repositories for an organization or a team.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  set_params(
    search_param = "team",
    team_name = "rwdevs"
  ) \%>\%
  set_team_member("Maciej Banaś", "maciekbanas") \%>\%
  pull_commits(date_from = "2018-01-01")
}
}
