% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{pull_repos_contributors}
\alias{pull_repos_contributors}
\title{Pull information on contributors}
\usage{
pull_repos_contributors(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} class object with repositories table with added
information on contributors.
}
\description{
Adds information on contributors to already pulled repositories
table.
}
\examples{
my_gitstats <- create_gitstats() \%>\%
  set_host(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma")
  ) \%>\%
  set_host(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "mbtests"
  ) \%>\%
  pull_repos() \%>\%
  pull_repos_contributors()
}
