% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{reset}
\alias{reset}
\title{Reset GitStats settings}
\usage{
reset(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} object.
}
\description{
Sets all settings to default.
}
