% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{reset_language}
\alias{reset_language}
\title{Reset language settings}
\usage{
reset_language(gitstats_obj)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}
}
\value{
A \code{GitStats} object.
}
\description{
Sets language parameter to \code{NULL} (switches of filtering by
language.).
}
