% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_connection.R
\name{set_connection}
\alias{set_connection}
\title{Setting connections}
\usage{
set_connection(gitstats_obj, api_url, token, orgs = NULL)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{api_url}{A character, url address of API.}

\item{token}{A token.}

\item{orgs}{A character vector of organisations (owners of repositories
in case of GitHub and groups of projects in case of GitLab).}
}
\value{
A \code{GitStats} class object with added information on connection
(\verb{$clients} field).
}
\description{
Setting connections
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_connection(
    api_url = "https://gitlab.com/api/v4",
    token = Sys.getenv("GITLAB_PAT_PUBLIC"),
    orgs = "erasmusmc-public-health"
  )
}
}
