% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{set_params}
\alias{set_params}
\title{Set up your search settings}
\usage{
set_params(
  gitstats_obj,
  search_param = NULL,
  team_name = NULL,
  phrase = NULL,
  files = NULL,
  language = "All",
  verbose = TRUE
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{search_param}{One of three: team, orgs or phrase.}

\item{team_name}{Name of a team.}

\item{phrase}{A phrase to look for.}

\item{files}{Define files to scan.}

\item{language}{Code programming language.}

\item{verbose}{A boolean to decide whether to print output.}
}
\value{
A \code{GitStats} object.
}
\description{
Set up your search settings
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_params(
    search_param = "team",
    team_name = "Avengers",
    language = "R"
  )
}
}
