% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{set_storage}
\alias{set_storage}
\title{Set a storage for your output}
\usage{
set_storage(
  gitstats_obj,
  type,
  dbname,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{type}{A character, type of storage.}

\item{dbname}{Name of database.}

\item{host}{A host.}

\item{port}{An integer, port address.}

\item{user}{Username.}

\item{password}{A password.}
}
\description{
Function to set local storage (database) for the output.
}
\details{
This is a wrapper over \code{dbConnect()} function. This functionality
is meant among other to improve \code{GitStats} performance. Basically the
idea is to cache your API responses in a database and retrieve them
when necessary. E.g. when you call \code{get_commits(date_from = '2020-01-01')} it will save the results to your database and when you
call it next time, \code{get_commits()} will retrieve from API only these
results that do not exist in the databse (the new ones).
}
