% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{set_team}
\alias{set_team}
\title{Set your team.}
\usage{
set_team(gitstats_obj, team_name, ...)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{team_name}{A name of a team.}

\item{...}{A charactger vector of team members (names and logins).}
}
\value{
A \code{GitStats} class object with added information to \verb{$team} field.
}
\description{
Declare your team members (logins, full names) to obtain
statistcis \code{by = "team"}.
}
\details{
Bear in mind that on different Git platforms, team members may use
different logins. You have to specify all of them, if you want to get team
statitistics from all your Git clients.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_connection(
    api_url = "https://api.github.com",
    token = Sys.getenv("GITHUB_PAT"),
    orgs = c("r-world-devs", "openpharma", "pharmaverse")
  ) \%>\%
  set_team("RWD-IE", "galachad", "kalimu", "Cotau", "krystian8207", "marcinkowskak", "maciekbanas") \%>\%
  get_repos(by = "team")
}
}
