% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{set_team_member}
\alias{set_team_member}
\title{Add your team member}
\usage{
set_team_member(gitstats_obj, member_name, ...)
}
\arguments{
\item{gitstats_obj}{\code{GitStats} object.}

\item{member_name}{Name of a member.}

\item{...}{All user logins.}
}
\value{
\code{GitStats} object with new information on team member.
}
\description{
Passes information on team member to your \code{team} field.
}
\examples{
\dontrun{
my_gitstats <- create_gitstats() \%>\%
  set_team_member("Peter Parker", "spider_man", "spidey") \%>\%
  set_team_member("Tony Stark", "ironMan", "tony_s")
}
}
