% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStats.R
\name{setup_preferences}
\alias{setup_preferences}
\title{Set up your search parameters.}
\usage{
setup_preferences(
  gitstats_obj,
  search_param = NULL,
  team = NULL,
  orgs = NULL,
  phrase = NULL,
  language = NULL
)
}
\arguments{
\item{gitstats_obj}{A GitStats object.}

\item{search_param}{One of three: team, orgs or phrase.}
}
\value{
A \code{GitStats} object.
}
\description{
Set up your search parameters.
}
