% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{show_orgs}
\alias{show_orgs}
\title{Show organizations set in \code{GitStats}}
\usage{
show_orgs(gitstats_object)
}
\arguments{
\item{gitstats_object}{A GitStats object.}
}
\value{
A vector of organizations.
}
\description{
Retrieves organizations set or pulled by \code{GitStats}. Especially
helpful when user is scanning whole git platform and wants to have a
glimpse at organizations.
}
