% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstats_functions.R
\name{verbose_on}
\alias{verbose_on}
\title{Switch on verbose mode}
\usage{
verbose_on(gitstats)
}
\arguments{
\item{gitstats}{A GitStats object.}
}
\value{
A GitStats object.
}
\description{
Print all messages and output.
}
